requirements.txt
selenium
python-dotenv
pyjq
tabulate